+++ src/main/java/com/ae2powertools/Tags.java
@@ -6,5 +6,5 @@ package com.ae2powertools;
 public class Tags {
     public static final String MODID = "ae2powertools";
     public static final String MODNAME = "AE2 PowerTools";
-    public static final String VERSION = "1.3.0";
+    public static final String VERSION = "1.3.2";
 }